Full-stack Test Kit Connection Update
Did few changes here and there. The log table has now an action column where the changes within the test kit records can be made (Price and Stock Count). The Put request from the front-end to back-end is not functioning as intended and requires edits. And additionally, created css for more welcoming experience.